pyproject.toml
[tool.poetry]
name = "zeigen"
version = "0.2.4"
description = "Find water networks in atomic-resolution crystal structures"
authors = ["Joel Berendzen <joel@generisbio.com>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/hydrationdynamics/zeigen"
repository = "https://github.com/hydrationdynamics/zeigen"
documentation = "https://zeigen.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Chemistry"
]
keywords = [
    "science",
    "chemistry",
    "physical chemistry",
    "biology",
    "structural biology",
    "hydrated water"
]
packages = [
	{ include = "zeigen", from = "src" },
	{ include = "rcsbsearch", from = "src" }
]

[tool.poetry.urls]
Changelog = "https://github.com/hydrationdynamics/zeigen/releases"

[tool.poetry.dependencies]
python = ">=3.9.1,<3.11"
attrs = ">=21.4.0"
biopython = ">=1.79"
colorama = ">=0.4.4"
dotli = ">=1.1"
dynaconf = ">=3.1.7"
gemmi = ">=0.5.2"
gql = {extras = ["all"], version = ">=3.0.0"}
loguru = ">=0.6.0"
matplotlib = ">=3.5.1"
numpy = ">=1.22.1"
pandas = ">=1.4.0"
pint = ">=0.18"
pyarrow = ">=9.0.0"
#rcsbsearch = {git = "https://github.com/hydrationdynamics/rcsbsearch.git", branch="master" }
schema = ">=0.7.5"
scipy = ">=1.7.3"
statsdict = ">=0.1.5"
tabulate = ">=0.8.9"
toml = ">=0.10.2"
tqdm = ">=4.64.1"
typer = ">=0.4.0"
uncertainties = ">=3.1.6"
jsonschema = "^4.16.0"
requests = "^2.28.1"
requests-download = "^0.1.2"
progressbar2 = "^4.0.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
loguru-mypy = ">=0.0.4"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
# test deps for this package
nox = ">=2021.6.6"
nox-poetry = ">=0.9.0"
pandas-stubs = ">=1.2.0.45"
pytest-cov = ">=3.0.0"
pytest-datadir-mgr = ">=1.3.1"
sh = ">=1.14.2"
types-requests = "^2.28.11.2"
types-tabulate = ">=0.8.5"
types-toml = ">=0.10.3"

[tool.poetry.scripts]
zeigen = "zeigen.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
source = ["zeigen"]

[tool.coverage.report]
show_missing = true

[tool.cruft]
skip = ["src/*", "tests/*", "poetry.lock"]

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
#warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
addopts = "-x"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
